pyproject.toml
[tool.poetry]
name = "javis-api"
version = "0.1.0"
description = ""
authors = ["Johnny Kim <bdkim855@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
openai = "^1.35.3"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.31"
fastapi-sqlalchemy = "^0.2.1"
alembic = "^1.13.1"
pymysql2 = "^1.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
